Added article to context
