--- client/blocks/inline-help/inline-help-center-types.ts
@@ -1,5 +1,0 @@
-export interface Article {
-	title: string;
-	link?: string;
-	icon?: string;
-}
